Add support for sets of disks
As you know, qcow2 images can be based off another image.  This patch
allows the creation of a new set of disks based on another set that
already exist.

So one can leave 'base' as a pristine Fedora install, and then have a
'puppet' set that has the puppet config applied.  One can then wipe
the puppet set as often as one likes and not have to wait for the
Fedora install again.