Ask the user about missing Fedora ISOs
I'm attempting to reduce the documentation required to run this
situation.  Aim is to be able to just run `./install` and it'll ask
the user any things it needs to know.

This patch makes it ask the user if they'd like to download the Fedora
ISO, and if they wouldn't, then it asks them if they've got it
somewhere else on their system already.