Enable passing additional arguments to puppet.
It's occasionally useful to be able to pass arguments to puppet,
for instance to make the output more verbose. Add a new argument
'--pp-args' whose value will be used on the command line at the
remote end.

Sample usage:

    ./config --pp-args="-dv"

This would enable both debug and verbose output from puppet. Note
that using '--pp-args="value"' style of argument is important to
avoid argparse at this end trying to interpret the arguments and
emit errors.

Change-Id: Ifc9ade801f138419520b6269a90ec2060fa83caa